refactor: move Track type into track_factory
